+++ lua/pi/ui/highlights.lua
@@ -7,6 +7,7 @@ M.CHAT_PROMPT_ATTENTION_WINHIGHLIGHT =
     "NormalFloat:PiFloat,FloatBorder:PiFloatBorder,FloatTitle:PiChatPromptFloatAttentionTitle"
 M.CHAT_PROMPT_BASH_WINHIGHLIGHT = "NormalFloat:PiFloat,FloatBorder:PiFloatBorder,FloatTitle:PiChatPromptFloatBashTitle"
 M.CHAT_ATTACHMENTS_WINHIGHLIGHT = "NormalFloat:PiFloat,FloatBorder:PiFloatBorder,FloatTitle:PiChatAttachmentsFloatTitle"
+M.SESSIONS_LIST_WINHIGHLIGHT = "NormalFloat:PiFloat,FloatBorder:PiFloatBorder,FloatTitle:PiSessionsListFloatTitle"
 M.DIFF_WINHIGHLIGHT = "WinBar:PiDiffWinbar,WinBarNC:PiDiffWinbar"
 
 --- Clear the Pi* groups we previously installed as defaults.
@@ -192,6 +193,13 @@ local function set_defaults()
     vim.api.nvim_set_hl(0, "PiStatusLineAttention", { default = true, fg = warning.fg, bold = true })
     vim.api.nvim_set_hl(0, "PiStatusLineWarning", { default = true, fg = warning.fg })
     vim.api.nvim_set_hl(0, "PiStatusLineError", { default = true, fg = diagnostic_error.fg })
+
+    vim.api.nvim_set_hl(0, "PiSessionsListTab", { default = true, fg = comment.fg, bold = true })
+    vim.api.nvim_set_hl(0, "PiSessionsListIdle", { default = true, fg = comment.fg })
+    vim.api.nvim_set_hl(0, "PiSessionsListCompacting", { default = true, fg = special.fg, bold = true })
+    vim.api.nvim_set_hl(0, "PiSessionsListExited", { default = true, fg = diagnostic_error.fg })
+    vim.api.nvim_set_hl(0, "PiSessionsListPending", { default = true, fg = comment.fg, italic = true })
+    vim.api.nvim_set_hl(0, "PiSessionsListFloatTitle", { default = true, fg = title.fg, bold = true })
 end
 
 function M.setup()